feat: banner를 모듈화하고, post 퍼블리싱
