+++ src/containers/Article/CommentForm/index.tsx
@@ -1,0 +1,12 @@
+import styled from "@emotion/styled"
+import React from "react"
+
+type Props = {}
+
+const CommentForm: React.FC<Props> = ({}) => {
+  return <StyledWrapper>CommentForm</StyledWrapper>
+}
+
+export default CommentForm
+
+const StyledWrapper = styled.div``
